Apply renaming to systematics
