+++ src/plotIt.cc
@@ -56,16 +56,6 @@ static Dummy foo;
 
 namespace plotIt {
 
-  std::string applyRenaming(const std::vector<RenameOp>& ops, const std::string input) {
-      std::string result = input;
-
-      for (const auto& op: ops) {
-          result = std::regex_replace(result, op.from, op.to, std::regex_constants::format_sed);
-      }
-
-      return result;
-  }
-
   plotIt::plotIt(const fs::path& outputPath):
     m_outputPath(outputPath) {
